Initial configuration

Define several parameters required to configure the rjMCMC sampler.

configure_rjMCMC(
  dat,
  model.select = TRUE,
  covariate.select = FALSE,
  proposal.mh = list(t.ij = 10, mu.i = 10, mu = 7, phi = 10, sigma = 10),
  proposal.rj = list(dd = 20, cov = 7),
  prior.covariates = c(0, 30),
  n.rep = 1000
)

Arguments

dat
Input BRS data. Must be an object of class brsdata.

model.select
Logical. If TRUE, between-model jumps will be allowed in the rjMCMC sampler. This argument can be set to FALSE to impose a species grouping of constrain and constrain parameter estimation accordingly.

covariate.select
Logical. Set to TRUE to allow contextual covariates to drop out of the model.

proposal.mh
Named list specifying the standard deviations of the proposal distributions used in the Metropolis-Hastings sampler.

proposal.rj
Named list specifying the standard deviations of the proposal distributions used in the reversible jump sampler. Must contain two elements: mu for response thresholds, and cov for covariates.

prior.covariates
Mean and standard deviation of the Normal prior placed on covariates. At present, the same prior is applied to all covariates.

n.rep
Number of replicate bootstrap datasets used for clustering (see Details).

Value

A list object of class rjconfig identical to the input brsdata object, with an additional config element composed of:

var	Empirical estimates of φ and σ

prop	Standard deviations of MCMC proposal distributions

clust	Outputs from the cluster analysis

boot	Species group assignments for each bootstrap dataset

mlist	Unique species group assignments

prior	Mean and standard deviation of the Normal priors placed on covariates

...	Other elements pulled from the brsdata object

Details

The function performs three actions:

Variance

Returns empirical estimates of between-whale (φ) and within-whale between-exposure (σ) variation, which are needed to generate starting values for the MCMC chains

Proposals

Defines the means and standard deviations of relevant proposal distributions and priors

Clustering

Performs a cluster analysis using n.rep bootstrap replicates of the input data, in order to parameterise between-model jumps. The analysis is run using Mclust on both (1) species means and (2) raw observations. These respectively provide estimates of the probability distributions of (1) unique groupings, and (2) numbers of groups, where the latter ranges from n = 1 when all species belong to the same group, to n = n.species when each species is treated individually.

See also

Examples

library(espresso)

# Import the example data, excluding species 
# with less than 5 observations
mydat <- read_data(file = NULL, min.N = 5)
summary(mydat)

#> 
#> ======================================================
#> DATA SUMMARY
#> ======================================================
#> 
#> Simulation: FALSE
#> Data file: example_brs
#> 
#> --------------------
#> OBSERVATIONS
#> --------------------
#> Right-censoring: TRUE 
#> Total: 58 
#> 
#> --------------------
#> SPECIES
#> --------------------
#> N =  2 
#> Excluded: Ba Gg Ha Md Oo Zc 
#> Species groupings: FALSE
#> 
#> # A tibble: 2 x 8
#>   species common_name N_ind N_trials censored  mean   min   max
#>   <chr>   <chr>       <int>    <int>    <int> <dbl> <dbl> <dbl>
#> 1 Bm      Blue whale     40       42       21  132.  80.7  177.
#> 2 Pm      Sperm whale    18       49       28  137. 116.   159.
#> 
#> --------------------
#> COVARIATES
#> --------------------
#> No covariates

# Configure the rjMCMC sampler
mydat.config <- configure_rjMCMC(dat = mydat,
                                 model.select = TRUE,
                                 covariate.select = FALSE,
                                 proposal.mh = list(t.ij = 10, mu.i = 10,
                                                    mu = 7, phi = 10, sigma = 10),
                                 proposal.rj = list(dd = 20, cov = 7),
                                 prior.covariates = c(0, 30),
                                 n.rep = 100)
summary(mydat.config)

#> 
#> ======================================================
#> DATA SUMMARY
#> ======================================================
#> 
#> Simulation: FALSE
#> Data file: example_brs
#> 
#> --------------------
#> OBSERVATIONS
#> --------------------
#> Right-censoring: TRUE 
#> Total: 58 
#> 
#> --------------------
#> SPECIES
#> --------------------
#> N =  2 
#> Excluded: Ba Gg Ha Md Oo Zc 
#> Species groupings: FALSE
#> 
#> # A tibble: 2 x 8
#>   species common_name N_ind N_trials censored  mean   min   max
#>   <chr>   <chr>       <int>    <int>    <int> <dbl> <dbl> <dbl>
#> 1 Bm      Blue whale     40       42       21  132.  80.7  177.
#> 2 Pm      Sperm whale    18       49       28  137. 116.   159.
#> 
#> --------------------
#> COVARIATES
#> --------------------
#> No covariates
#> 
#> --------------------
#> MCMC
#> --------------------
#> 
#> Model selection: TRUE 
#> Covariate selection: FALSE 
#> 
#> Variance estimates: 5.7 (sigma); 14 (phi) 
#> 
#> Proposal standard deviations:
#> 
#> # A tibble: 6 x 3
#>   param          SD step 
#>   <chr>       <dbl> <chr>
#> 1 t.ij           10 MH   
#> 2 mu.i           10 MH   
#> 3 mu              7 MH   
#> 4 phi            10 MH   
#> 5 sigma          10 MH   
#> 6 data-driven    20 RJ   
#> 
#> Clustering:
#> 
#> [[1]]
#> # A tibble: 1 x 3
#>   model       p p_scale
#>   <chr>   <dbl>   <dbl>
#> 1 (Bm,Pm)     1       1
#> 
#> [[2]]
#> # A tibble: 2 x 2
#>   cluster     p
#>     <dbl> <dbl>
#> 1       1  0.25
#> 2       2  0.75
#>