Example dataset

Example dataset used to illustrate the format expected by espresso.

data(example_brs)

Format

A tibble, with the following columns:

Project	Project name or code

Species	Species code, as listed in species_brs

Tag_ID	Unique tag identifier

Resp_time	Time at response

Resp_type	Response type (avoidance, startle, respiration, foraging, locomotion, group, dive, orientation, resting, vocal, or any)

Resp_score	Severity of the response (on a point scale from 0 to 7), as scored by expert elicitation

Resp_SPL	Sound pressure level at time of response (in dB re 1μPa)

Resp_SELcum	Cumulative sound exposure level at time of response (in dB re 1μPa)

Max_SPL	Maximum sound pressure level reached during the exposure

Max_SELcum	Maximum sound exposure level reached during the exposure

Censored	Integer variable indicating whether an observation is left-censored (-1), right-censored (1) or not censored (0)

Exp_order	History of exposure (1 = 1st exposure, 2 = 2nd exposure, etc.)

Exp_duration	Duration of the exposure

Exp_signal	Type of sonar signal (e.g., MFAS, REAL MFA, PRN, CAS)

Pre_feeding	Behavioural mode (TRUE = feeding, FALSE = non-feeding)

Resp_range	Whale-source range at the time of response

Min_range	Minimum whale-source range during the exposure

Inferred Resp_range	Best estimate of whale-source range at the time of response

Inferred Min_range	Best estimate of minimum whale-source range during the exposure

Details

Note: This is a completely manufactured dataset, which should not be used for inference. The ordering of columns does not matter.