Dose-response curves

Plot dose_response curves. objects produced by compile_rjMCMC.

# S3 method for dose_response
plot(
  dr.object,
  model.rank = 1,
  overlay = FALSE,
  colour.by = NULL,
  colour = NULL,
  colour.median = NULL,
  order.by = "species",
  show.p0_5 = TRUE,
  rotate.y = FALSE,
  all.credint = FALSE,
  scientific.name = FALSE,
  common.name = FALSE,
  overwrite.abbrev = TRUE,
  outline.outer = FALSE
)

Arguments

dr.object
Input object of class dose_response, as returned by compile_rjMCMC.

model.rank
Integer indicating the rank of the model for which dose-response curves are to be plotted. The model with highest posterior probability is given a rank of 1.

overlay
Logical. If TRUE, plots will be faceted by species (or species group). If FALSE, all curves will be overlaid on the same plot.

colour.by
Curves can be coloured by species group (colour.by = "group") or by individual species (colour.by = "species"), or can use a single colour scheme (the default, colour.by = NULL).

colour
Used to overwrite the colour.by argument, if desired.

colour.median
Colour to use for the posterior median. Overwrites the colour.by argument.

order.by
How should plots be arranged? Use order.by = "species" to arrange plots by species (groups) names in alphabetical order, or order.by = "response", to arrange plots by response thresholds (from most to least sensitive species (groups), as determined by the estimated posterior medians for μ). Only relevant when overlay = FALSE.

show.p0_5
Logical. If TRUE, the values of posterior medians are added to facet labels. Only relevant when overlay = FALSE.

rotate.y
Logical. If TRUE, y-axis labels are rotated clockwise by 90 degrees.

all.credint
Logical. If TRUE, plot all credible intervals. If FALSE, only plot the outermost interval.

scientific.name
Logical. If TRUE, use species' scientific names in plot titles.

common.name
Logical. If TRUE, use species' common names in plot titles.

overwrite.abbrev
Logical. If TRUE, overwrites abbreviated names for species grouped a priori using create_groups.

outline.outer
Logical. If TRUE, outline the outermost credible interval bounds.

See also

Author

Phil J. Bouchet

Examples

if (FALSE) {
library(espresso)

# Import the example data, excluding species with sample sizes < 5
# and considering the sonar covariate
mydat <- read_data(file = NULL, min.N = 5, covariates = "sonar") 
summary(mydat)

# Configure the sampler
mydat.config <- configure_rjMCMC(dat = mydat,
                                 model.select = TRUE,
                                 covariate.select = FALSE,
                                 proposal.mh = list(t.ij = 10, mu.i = 10, 
                                                    mu = 7, phi = 10, sigma = 10),
                                 proposal.rj = list(dd = 20, cov = 7),
                                 prior.covariates = c(0, 30),
                                 n.rep = 100)
summary(mydat.config)

# Run the reversible jump MCMC
rj <- run_rjMCMC(dat = mydat.config,
                 n.chains = 2,
                 n.burn = 100,
                 n.iter = 100,
                 do.update = FALSE)
# Burn and thin
rj.trace <- trace_rjMCMC(rj.dat = rj)

# Get dose-response functions
doseR <- compile_rjMCMC(rj.trace)

# Plot dose-response curves
plot(dr.object = doseR, colour.by = "group")
}