Diagnostic plots

Generate trace, density, and autocorrelation plots from an rjtrace object.

# S3 method for rjtrace
plot(
  rj.obj,
  param.name = "all",
  covariates.incl = TRUE,
  autocorr = FALSE,
  individual = TRUE
)

Arguments

rj.obj
rjMCMC trace object of class rjtrace.

param.name
Parameter name(s). Defaults to all, which returns plots for all parameters in the model.

covariates.incl
Logical. If TRUE, the trace is filtered to only retain posterior estimates obtained when the contextual covariates were included in the model. Only relevant when covariate.select = TRUE in configure_rjMCMC.

autocorr
Logical. Whether to output chain autocorrelation plots.

individual
Logical. If TRUE, separate density lines will be plotted for each chain. If FALSE, one density line will be plotted for all chains.

Details

Adapted from Casey Youngflesh's MCMCtrace.

See also

Examples

library(espresso)

# Import the example data

mydat <- read_data(file = NULL)

#> Warning: N < 5 for some species. Species with low sample sizes may need to be grouped a priori.

# Import a real dataset with the sonar and range covariates, 
# excluding sperm whales and any other species with a sample size
# smaller than two

mydat <- read_data(file = "path/to/my/data.csv",
                  exclude.species = "Sperm whale",
                  min.N = 2)

#> Error: 'path/to/my/data.csv' does not exist in current working directory ('/Users/philippebouchet/Google Drive/Documents/git/espresso/docs/reference').