Set up the rjMCMC sampler

Set up a list object to hold the data from the MCMC sampler, and generate starting values for all model parameters. This function is called internally by run_rjMCMC.

setup_rjMCMC(
  rj.input,
  n.burn,
  n.iter,
  n.chains = 3,
  p.split = 0.5,
  p.merge = 0.5,
  move.ratio = list(dd1 = 3, dd2 = 1, random = 1),
  m = 100,
  do.update = FALSE,
  start.values = NULL
)

Arguments

rj.input
Input dataset. Must be an object of class rjconfig.

n.burn
Number of MCMC iterations to treat as burn-in.

n.iter
Number of posterior samples.

n.chains
Number of MCMC chains. Defaults to 3.

p.split
Probability of choosing to split a group of species when initiating a split/merge move. This parameter is constrained to be 1 when all species are in a single group. p.split and p.merge must sum to 1.

p.merge
Probability of choosing to merge two groups of species when initiating a split/merge move. This parameter is constrained to be 1 when all species are in their own groups. p.split and p.merge must sum to 1.

move.ratio
Relative proportion of calls to data-driven (type I and II) and independence samplers.

m
Integer. Frequency (every m iterations) at which data-driven (type I and II) and independence samplers are triggered.

do.update
Logical. Whether to update an existing sampler or set up a new one.

start.values
Starting values to use when updating an existing sampler and do.update = TRUE.

Details

In addition to split/merge moves, three other types of MCMC samplers are implemented in espresso to accelerate convergence and avoid getting stuck in local maxima: two data-driven samplers (type I and type II), in which proposals are informed by the “cues” present in the original data, and one independent sampler, in which proposals are drawn at random from a Uniform distribution bounded by range.dB(see read_data or simulate_data), with no dependence on current values.

See also