Some diagnostics for fitted dose-response models

Summary method for objects of class gvs, as returned by gibbs. Produces a text-based summary of: (1) effective sample sizes, (2) acceptance rates, (3) model convergence, and (4) posterior model probabilities. Note that gibbs does not implement covariate selection. As a result, posterior inclusion probabilities (PIPs) are not returned here, contrary to summary.rjtrace.

# S3 method for gvs
summary(
  gvs.obj,
  eff.n = TRUE,
  accept.rate = TRUE,
  convergence = TRUE,
  gelman.rubin = 1.1,
  model.ranks = TRUE,
  n.top = NULL
)

Arguments

gvs.obj
Input trace object, as returned by gibbs.

eff.n
Logical. If TRUE, returns estimates of the effective sample size for each model parameter.

accept.rate
Logical. If TRUE, returns the acceptance rate (calculated after burn-in) for each model parameter.

convergence
Logical. If TRUE, assesses convergence using the multivariate potential scale reduction factor (Gelman-Rubin statistic), as implemented in gelman.diag.

gelman.rubin
Threshold for determining convergence based on the Gelman-Rubin statistic. Defaults to 1.1.

model.ranks
Logical. If TRUE, returns a summary of posterior model probabilities and associated model rankings.

n.top
Number of top-ranking models to display when model.ranks = TRUE. Defaults to NULL, which lists all models.

Value

A detailed summary, printed to the R console.

See also

Examples

library(espresso)

# Simulate data for two species
mydat <- simulate_data(n.species = 2,
                       n.whales = 16,
                       max.trials = 3,
                       covariates = list(exposed = c(0, 5), range = 0.5),
                       mu = c(101, 158),
                       phi = 20,
                       sigma = 20,
                       Rc = c(210, 211),
                       seed = 58697)
summary(mydat)

#> 
#> ======================================================
#> DATA SUMMARY
#> ======================================================
#> 
#> Simulation: TRUE
#> 
#> --------------------
#> OBSERVATIONS
#> --------------------
#> Right-censoring: TRUE 
#> Total: 1 
#> 
#> --------------------
#> SPECIES
#> --------------------
#> N =  2 
#> Excluded: None 
#> Species groupings: FALSE
#> 
#> # A tibble: 2 x 6
#>   species N_ind N_trials  mean   min   max
#>     <dbl> <dbl>    <int> <dbl> <dbl> <dbl>
#> 1       1     8       12  116.  60.4  172.
#> 2       2     8       17  162.  98.3  210.
#> 
#> --------------------
#> COVARIATES
#> --------------------
#> --  exposed  --
#> Levels: 2 
#> [0]: n = 16
#> [1]: n = 13
#> [NA]: n = 0
#> 
#> --  range  --
#> [min]: 0.068
#> [mean]: 10.127
#> [max]: 19.431
#> [NA]: n = 0
#> 
#> 

# Model selection by GVS
gvs.model <- gibbs(dat = mydat,
             random.effects = FALSE,
             include.covariates = FALSE,
             mcmc.n = 1000,
             burnin = 500)

#> 
#> --------------------------------------------------
#> PSEUDO-PRIORS
#> --------------------------------------------------
#> 
#>  Candidate model (1 out of 2): (1)+(2)
#> 
#>  Candidate model (2 out of 2): (1,2)
#> 
#> --------------------------------------------------
#> MAIN MODEL
#> --------------------------------------------------
#> 

summary(gvs.model)

#> 
#> ======================================================
#> GVS SUMMARY
#> ======================================================
#> 
#> Iterations: 501:1500 
#> Thinning interval: 1 
#> Burn-in: 500 
#> Number of chains: 1 
#> Sample size per chain: 1,000 
#> Total sample size: 1,000 
#> 
#> Run times:
#> # A tibble: 1 x 2
#>   Chains Time  
#>   <chr>  <time>
#> 1 1:1    00'07"
#> 
#> --------------------
#> EFFECTIVE SAMPLE SIZES
#> --------------------
#>  mu_i[1]  mu_i[2]      phi    sigma    theta 
#> 61.96209 97.49142 28.76053 91.83114 39.43465 
#> 
#> --------------------
#> ACCEPTANCE RATES
#> --------------------
#> mu_i[1] mu_i[2]     phi   sigma   theta 
#>   1.000   1.000   1.000   1.000   0.024 
#> 
#> --------------------
#> MODEL RANKINGS
#> --------------------
#> (Chains 1:1):
#> # A tibble: 2 x 4
#>   model       p lower upper
#>   <fct>   <dbl> <dbl> <dbl>
#> 1 (1)+(2) 0.924 0.840 1    
#> 2 (1,2)   0.076 0     0.160
#>