Some diagnostics for fitted dose-response models

Summary method for objects of class rjtrace, as returned by trace_rjMCMC. Produces a text-based summary of: (1) effective sample sizes, (2) acceptance rates, (3) model convergence, (4) posterior model probabilities, and (5) posterior inclusion probabilities (PIPs) for contextual covariates (where appripriate).

# S3 method for rjtrace
summary(
  rj.obj,
  eff.n = TRUE,
  accept.rate = TRUE,
  convergence = TRUE,
  gelman.rubin = 1.1,
  model.ranks = TRUE,
  n.top = NULL,
  covariate.prob = TRUE,
  combine.chains = FALSE
)

Arguments

rj.obj
Input rjMCMC object, as returned by trace_rjMCMC.

covariate.prob
Logical. If TRUE, returns a summary of posterior inclusion probabilities (PIPs).

combine.chains
Logical. If TRUE, outputs are returned for each individual MCMC chain.

Value

A detailed summary, printed to the R console.

See also

Examples

library(espresso)

# Import the example data, excluding species with sample sizes < 5
# and considering the sonar covariate
mydat <- read_data(file = NULL, min.N = 5, covariates = "sonar")
summary(mydat)

#> 
#> ======================================================
#> DATA SUMMARY
#> ======================================================
#> 
#> Simulation: FALSE
#> Data file: example_brs
#> 
#> --------------------
#> OBSERVATIONS
#> --------------------
#> Right-censoring: TRUE 
#> Total: 46 
#> 
#> --------------------
#> SPECIES
#> --------------------
#> N =  2 
#> Excluded: Ba Ha Md Oo Zc 
#> Species groupings: FALSE
#> 
#> # A tibble: 2 x 8
#>   species common_name N_ind N_trials censored  mean   min   max
#>   <chr>   <chr>       <int>    <int>    <int> <dbl> <dbl> <dbl>
#> 1 Bm      Blue whale     31       33       14  133.  80.7  177.
#> 2 Pm      Sperm whale    18       38       23  139. 123.   159.
#> 
#> --------------------
#> COVARIATES
#> --------------------
#> --  sonar  --
#> Levels: 2 
#> [LFAS]: n = 35
#> [MFAS]: n = 36
#> [NA]: n = 0
#> 
#> 
#> --------------------
#> SONAR GROUPINGS
#> --------------------
#> $MFAS
#> [1] "MFAS"          "MFA"           "REAL MFA"      "MFAS_DS"      
#> [5] "MFA HELO"      "REAL MFA HELO" "SOCAL_d"       "REAL MFAS"    
#> [9] "MF ALARM"     
#> 
#> $LFAS
#> [1] "LFAS_LO"  "LFAS"     "LFAS_DS"  "LFAS_122" "LFAS_185"
#> 
#> $exclude
#> [1] "PRN"          "CAS"          "ALARM"        "HF ALARM"     "HFAS"        
#> [6] "HF ALARM CAS" "MFAS CAS"    
#> 

# Configure the sampler
mydat.config <- configure_rjMCMC(dat = mydat,
                                 model.select = TRUE,
                                 covariate.select = FALSE,
                                 proposal.mh = list(t.ij = 10, mu.i = 10,
                                                    mu = 7, phi = 10, sigma = 10),
                                 proposal.rj = list(dd = 20, cov = 7),
                                 prior.covariates = c(0, 30),
                                 n.rep = 100)
summary(mydat.config)

#> 
#> ======================================================
#> DATA SUMMARY
#> ======================================================
#> 
#> Simulation: FALSE
#> Data file: example_brs
#> 
#> --------------------
#> OBSERVATIONS
#> --------------------
#> Right-censoring: TRUE 
#> Total: 46 
#> 
#> --------------------
#> SPECIES
#> --------------------
#> N =  2 
#> Excluded: Ba Ha Md Oo Zc 
#> Species groupings: FALSE
#> 
#> # A tibble: 2 x 8
#>   species common_name N_ind N_trials censored  mean   min   max
#>   <chr>   <chr>       <int>    <int>    <int> <dbl> <dbl> <dbl>
#> 1 Bm      Blue whale     31       33       14  133.  80.7  177.
#> 2 Pm      Sperm whale    18       38       23  139. 123.   159.
#> 
#> --------------------
#> COVARIATES
#> --------------------
#> --  sonar  --
#> Levels: 2 
#> [LFAS]: n = 35
#> [MFAS]: n = 36
#> [NA]: n = 0
#> 
#> 
#> --------------------
#> SONAR GROUPINGS
#> --------------------
#> $MFAS
#> [1] "MFAS"          "MFA"           "REAL MFA"      "MFAS_DS"      
#> [5] "MFA HELO"      "REAL MFA HELO" "SOCAL_d"       "REAL MFAS"    
#> [9] "MF ALARM"     
#> 
#> $LFAS
#> [1] "LFAS_LO"  "LFAS"     "LFAS_DS"  "LFAS_122" "LFAS_185"
#> 
#> $exclude
#> [1] "PRN"          "CAS"          "ALARM"        "HF ALARM"     "HFAS"        
#> [6] "HF ALARM CAS" "MFAS CAS"    
#> 
#> 
#> --------------------
#> MCMC
#> --------------------
#> 
#> Model selection: TRUE 
#> Covariate selection: FALSE 
#> 
#> Variance estimates: 2.4 (sigma); 14.1 (phi) 
#> 
#> Proposal standard deviations:
#> 
#> # A tibble: 7 x 3
#>   param          SD step 
#>   <chr>       <dbl> <chr>
#> 1 t.ij           10 MH   
#> 2 mu.i           10 MH   
#> 3 mu              7 MH   
#> 4 phi            10 MH   
#> 5 sigma          10 MH   
#> 6 sonar          10 MH   
#> 7 data-driven    20 RJ   
#> 
#> Clustering:
#> 
#> [[1]]
#> # A tibble: 1 x 3
#>   model       p p_scale
#>   <chr>   <dbl>   <dbl>
#> 1 (Bm,Pm)     1       1
#> 
#> [[2]]
#> # A tibble: 2 x 2
#>   cluster     p
#>     <dbl> <dbl>
#> 1       1  0.27
#> 2       2  0.73
#> 

# Run the reversible jump MCMC
rj <- run_rjMCMC(dat = mydat.config,
                 n.chains = 2,
                 n.burn = 100,
                 n.iter = 100,
                 do.update = FALSE)
summary(rj)

#>      Length Class  Mode
#> [1,] 18     -none- list
#> [2,] 18     -none- list