Dissolve groupings

Undo species groupings generated by create_groups

undo_groups(dat.obj)

Arguments

dat.obj
Input dataset. Must be an object of class brsdata.group, as given by create_groups.

Value

A list object of class brsdata.

See also

Examples

library(espresso)

# Import the example data
mydat <- read_data(file = NULL)

#> Warning: N < 5 for some species. Species with low sample sizes may need to be grouped a priori.

summary(mydat)

#> 
#> ======================================================
#> DATA SUMMARY
#> ======================================================
#> 
#> Simulation: FALSE
#> Data file: example_brs
#> 
#> --------------------
#> OBSERVATIONS
#> --------------------
#> Right-censoring: TRUE 
#> Total: 70 
#> 
#> --------------------
#> SPECIES
#> --------------------
#> N =  8 
#> Excluded: None 
#> Species groupings: FALSE
#> 
#> # A tibble: 8 x 8
#>   species common_name               N_ind N_trials censored  mean   min    max
#>   <chr>   <chr>                     <int>    <int>    <int> <dbl> <dbl>  <dbl>
#> 1 Md      Blainville's beaked whale     2        2        2 NaN   Inf   -Inf  
#> 2 Bm      Blue whale                   40       42       21 132.   80.7  177. 
#> 3 Zc      Cuvier's beaked whale         3        3        1 120.  112.   129. 
#> 4 Oo      Killer whale                  4        8        5 113.  101.   126. 
#> 5 Ba      Minke whale                   2        2        1 114.  114.   114. 
#> 6 Ha      Northern bottlenose whale     4        4        3 144.  144.   144. 
#> 7 Gg      Risso's dolphin               1        1        0  99.3  99.3   99.3
#> 8 Pm      Sperm whale                  18       49       28 137.  116.   159. 
#> 
#> --------------------
#> COVARIATES
#> --------------------
#> No covariates

# Group all beaked whales together
mydat.grouped <- create_groups(mydat, species.groups = list(beaked = c("Md", "Zc", "Ha")))
class(mydat.grouped)

#> [1] "brsdata.grp" "brsdata"     "list"       

summary(mydat.grouped)

#> 
#> ======================================================
#> DATA SUMMARY
#> ======================================================
#> 
#> Simulation: FALSE
#> Data file: example_brs
#> 
#> --------------------
#> OBSERVATIONS
#> --------------------
#> Right-censoring: TRUE 
#> Total: 70 
#> 
#> --------------------
#> SPECIES
#> --------------------
#> N =  6 
#> Excluded: None 
#> Species groupings:
#> 
#> # A tibble: 1 x 2
#>   name   species 
#>   <chr>  <chr>   
#> 1 beaked Md,Zc,Ha
#> 
#> # A tibble: 6 x 7
#>   species N_ind N_trials censored  mean   min   max
#>   <chr>   <int>    <int>    <int> <dbl> <dbl> <dbl>
#> 1 Ba          2        2        1 114.  114.  114. 
#> 2 beaked      9        9        6 128.  112.  144. 
#> 3 Bm         40       42       21 132.   80.7 177. 
#> 4 Gg          1        1        0  99.3  99.3  99.3
#> 5 Oo          4        8        5 113.  101.  126. 
#> 6 Pm         18       49       28 137.  116.  159. 
#> 
#> --------------------
#> COVARIATES
#> --------------------
#> No covariates

# Undo groupings and revert back to the original data
mydat.ungrouped <- undo_groups(mydat.grouped)
class(mydat.ungrouped)

#> [1] "brsdata" "list"   

summary(mydat.ungrouped)

#> 
#> ======================================================
#> DATA SUMMARY
#> ======================================================
#> 
#> Simulation: FALSE
#> Data file: example_brs
#> 
#> --------------------
#> OBSERVATIONS
#> --------------------
#> Right-censoring: TRUE 
#> Total: 70 
#> 
#> --------------------
#> SPECIES
#> --------------------
#> N =  8 
#> Excluded: None 
#> Species groupings: FALSE
#> 
#> # A tibble: 8 x 8
#>   species common_name               N_ind N_trials censored  mean   min    max
#>   <chr>   <chr>                     <int>    <int>    <int> <dbl> <dbl>  <dbl>
#> 1 Md      Blainville's beaked whale     2        2        2 NaN   Inf   -Inf  
#> 2 Bm      Blue whale                   40       42       21 132.   80.7  177. 
#> 3 Zc      Cuvier's beaked whale         3        3        1 120.  112.   129. 
#> 4 Oo      Killer whale                  4        8        5 113.  101.   126. 
#> 5 Ba      Minke whale                   2        2        1 114.  114.   114. 
#> 6 Ha      Northern bottlenose whale     4        4        3 144.  144.   144. 
#> 7 Gg      Risso's dolphin               1        1        0  99.3  99.3   99.3
#> 8 Pm      Sperm whale                  18       49       28 137.  116.   159. 
#> 
#> --------------------
#> COVARIATES
#> --------------------
#> No covariates