Initial configuration

Define several parameters required to configure the rjMCMC sampler.

configure_rjMCMC(
  dat,
  model.select = TRUE,
  covariate.select = FALSE,
  proposal.mh = list(t.ij = 10, mu.i = 10, mu = 7, phi = 10, sigma = 10),
  proposal.rj = list(dd = 1, cov = 7),
  prior.covariates = c(0, 30),
  p.split = 0.5,
  p.merge = 0.5,
  moves = 0:3,
  move.ratio = list(dd1 = 3, dd2 = 1, random = 1),
  m = 100,
  bootstrap = TRUE,
  n.rep = 1000
)

Arguments

dat
Input BRS data. Must be an object of class brsdata.

model.select
Logical. If TRUE, between-model jumps will be allowed in the rjMCMC sampler. This argument can be set to FALSE to impose a species grouping and constrain parameter estimation accordingly.

covariate.select
Logical. Set to TRUE to allow contextual covariates to drop out of the model.

proposal.mh
Named list specifying the standard deviations of the proposal distributions used in the Metropolis-Hastings sampler.

proposal.rj
Named list specifying the standard deviations of the proposal distributions used in the reversible jump sampler. Must contain two elements: mu for response thresholds, and cov for covariates.

prior.covariates
Mean and standard deviation of the Normal prior placed on covariates. At present, the same prior is applied to all covariates.

p.split
Probability of choosing to split a group of species when initiating a split-merge move. This parameter is constrained to be 1 when all species are in a single group. p.split and p.merge must sum to 1.

p.merge
Probability of choosing to merge two groups of species when initiating a split-merge move. This parameter is constrained to be 1 when all species are in their own groups. p.split and p.merge must sum to 1.

moves
Integer indicating which between-model moves to implement. 0 = split/merge, 1 = data driven Type I, 2 = data driven Type II, 3 = random. Useful for testing purposes.

move.ratio
Relative proportion of calls to data-driven (type I and II) and independence samplers.

m
Integer. Frequency (every m iterations) at which data-driven (type I and II) and independence samplers are triggered.

bootstrap
Logical, defaults to TRUE. Whether to perform bootstrap clustering. As this step can be time-consuming, setting this argument to FALSE increase efficiency when reconfiguring the sampler.

n.rep
Number of replicate bootstrap datasets used for clustering (see Details).

Value

A list object of class rjconfig identical to the input brsdata object, with an additional config element composed of:

var	Empirical estimates of φ and σ

prop	Standard deviations of MCMC proposal distributions

clust	Outputs from the cluster analysis

boot	Species group assignments for each bootstrap dataset

mlist	Unique species group assignments

prior	Mean and standard deviation of the Normal priors placed on covariates

...	Other elements pulled from the brsdata object

Details

The function performs three actions:

Variance

Returns empirical estimates of between-whale (φ) and within-whale between-exposure (σ) variation, which are needed to generate starting values for the MCMC chains

Proposals

Defines the means and standard deviations of relevant proposal distributions and priors

Clustering

Performs a cluster analysis using n.rep bootstrap replicates of the input data, in order to parameterise between-model jumps. The analysis is run using Mclust on both (1) species means and (2) raw observations. These respectively provide estimates of the probability distributions of (1) unique groupings, and (2) numbers of groups, where the latter ranges from n = 1 when all species belong to the same group, to n = n.species when each species is treated individually.

See also

Author

Phil J. Bouchet

Examples

if (FALSE) {
library(espresso)

# Import the example data, excluding species 
# with less than 5 observations
mydat <- read_data(file = NULL, min.N = 5) 
summary(mydat)

# Configure the rjMCMC sampler
mydat.config <- configure_rjMCMC(dat = mydat,
                                 model.select = TRUE,
                                 covariate.select = FALSE,
                                 proposal.mh = list(t.ij = 10, mu.i = 10, 
                                                    mu = 7, phi = 10, sigma = 10),
                                 proposal.rj = list(dd = 20, cov = 7),
                                 prior.covariates = c(0, 30),
                                 n.rep = 100)
summary(mydat.config)
}