Dose-response curves

Plot method for dose_response objects produced by compile_rjMCMC.

# S3 method for dose_response
plot(
  dr.object,
  which.model = 1,
  overlay = FALSE,
  colour.by = NULL,
  colour = NULL,
  colour.median = NULL,
  order.by = "species",
  show.p0_5 = TRUE,
  rotate.y = FALSE,
  all.quants = FALSE,
  scientific.name = TRUE,
  common.name = FALSE,
  overwrite.abbrev = TRUE,
  outline.95 = FALSE
)

Arguments

dr.object
Input object of class dose_response.

which.model
Integer indicating the ID of the model for which dose-response curves are to be plotted. The summary.dose_response function can be used to list model IDs, if necessary.

overlay
Logical. If TRUE, plots will be faceted by species (or species group). If FALSE, all curves will be overlaid on the same plot.

colour.by
By default, all dose-response curves follow the same colour scheme (colour.by = NULL). However, curves can also be coloured by group or by species.

colour
Used to overwrite the colour.by argument, if desired.

colour.median
Colour to use for the posterior median line. Overwrites the colour.by argument when specified.

order.by
How should plots be ordered? When order.by = "species", plots are ordered alphabetically. order.by = "response", plots are ordered by response thresholds (from most to least sensitive species, as determined by posterior medians for μ. Only relevant when overlay = FALSE.

show.p0_5
Logical. If TRUE, posterior medians are added to facet labels. Only relevant when overlay = FALSE.

rotate.y
Logical. If TRUE, y-axis labels are rotated clockwise by 90 degrees.

all.quants

scientific.name
Logical. If TRUE, use species' scientific names in plot titles.

common.name
Logical. If TRUE, use species' common names in plot titles.

overwrite.abbrev
Logical. If TRUE, overwrites abbreviated names for species grouped a priori using cerate_groups.

outline.95
Logical. If TRUE, outlines the 95% posterior credible interval bounds.

See also

Examples

library(espresso)

# Import the example data

mydat <- read_data(file = NULL)

#> Warning: N < 5 for some species. Species with low sample sizes may need to be grouped a priori.

# Import a real dataset with the sonar and range covariates, 
# excluding sperm whales and any other species with a sample size
# smaller than two

mydat <- read_data(file = "path/to/my/data.csv",
                  exclude.species = "Sperm whale",
                  min.N = 2)

#> Error: 'path/to/my/data.csv' does not exist in current working directory ('/Users/philippebouchet/Google Drive/Documents/git/espresso/docs/reference').