Generate a simulated dataset

Simulate cetacean responses to Navy sonar exposure, as in behavioural response studies (BRSs).

simulate_data(
  n.species = 2,
  n.whales = 20,
  max.trials = 3,
  covariates = NULL,
  mu = NULL,
  phi = NULL,
  sigma = NULL,
  range.dB = c(60, 215),
  range.var = c(0, 45),
  Rc = c(190, 195),
  obs.sd = 2.5,
  verbose = TRUE,
  seed = NULL
)

Arguments

n.species
Number of species.

n.whales
Number of individuals. Can be specified as a vector of length n.species, allowing for uneven samples sizes across species, or as an integer multiple of n.species for equal sample sizes across species.

max.trials
Maximum number of exposures per individual. Each animal is exposed x times, with x taken as a random draw from a Uniform distribution between 1 and max.trials.

covariates
Contextual covariates and their associated coefficients. Must be supplied as a named list, in which the baseline levels of factor covariates are given a coefficient of zero.

mu
Mean response threshold(s) for each species.

phi
Between-whale variance in response thresholds.

sigma
Within-whale between-exposure variance in response thresholds.

range.dB
Bounds for the dose-response function. Must be a vector of length 2. Defaults to: (1) a lower bound of 60 dB re 1μPa, taken as a conservative lower limit of detectability given hearing sensitivity and the lowest average sea noise conditions; and (2) an upper bound of 215 dB re 1μPa at/above which all animals are expected to respond. This upper bound is consistent with the maximum source levels employed in behavioural response studies (BRSs) to date.

range.var
Permissible range for the variance term(s) phi (between-whale variance) and sigma (between-exposure variance). Must be a vector of length 2.

Rc
Right-censoring interval. Values of the maximum realised dose for each exposure are generated as random draws from a Uniform distribution within these bounds, such that max.dose <- runif(n = n.trials, min = Rc[1], max = Rc[2]).

obs.sd
Measurement uncertainty (expressed as a standard deviation in received levels), in dB re 1μPa.

verbose
Logical. Whether to print or suppress warning messages.

seed
Random seed (for reproducible results).

Value

A list object of class brsdata, with the following elements:

dat	Output dataset, after processing.

species	Species data, including species names, groups, sample sizes etc.

whales	Individual data, including animal IDs.

trials	Exposure data, including exposure IDs.

covariates	Covariate data, including dummy coding, sonar groupings, factor levels etc.

obs	Observations, including right-censoring cutoffs.

param	General parameters.

See also

Examples

library(espresso)

# Simulate data for two species
mydat <- simulate_data(n.species = 2,
                       n.whales = 16,
                       max.trials = 3,
                       covariates = list(exposed = c(0, 5), range = 0.5),
                       mu = c(125, 158),
                       phi = 20,
                       sigma = 20,
                       Rc = c(210, 211),
                       seed = 58697)