Some diagnostics for fitted dose-response models

Summary method for objects of class rjtrace, as returned by trace_rjMCMC. Produces a text-based summary of: (1) effective sample sizes, (2) acceptance rates, (3) model convergence, (4) posterior model probabilities, and (5) posterior inclusion probabilities (PIPs) for contextual covariates (where appripriate).

# S3 method for rjtrace
summary(
  rj.obj,
  eff.n = TRUE,
  accept.rate = TRUE,
  convergence = TRUE,
  gelman.rubin = 1.1,
  model.ranks = TRUE,
  n.top = NULL,
  covariate.prob = TRUE,
  combine.chains = FALSE
)

Arguments

rj.obj
Input rjMCMC object, as returned by trace_rjMCMC.

eff.n
Logical. If TRUE, returns estimates of the effective sample size for each model parameter.

accept.rate
Logical. If TRUE, returns the acceptance rate (calculated after burn-in) for each model parameter.

convergence
Logical. If TRUE, assesses convergence using the multivariate potential scale reduction factor (Gelman-Rubin statistic), as implemented in gelman.diag.

gelman.rubin
Threshold for determining convergence based on the Gelman-Rubin statistic. Defaults to 1.1.

model.ranks
Logical. If TRUE, returns a summary of posterior model probabilities and associated model rankings.

n.top
Number of top-ranking models to display when model.ranks = TRUE. Defaults to NULL, which lists all models.

covariate.prob
Logical. If TRUE, returns a summary of posterior inclusion probabilities (PIPs).

combine.chains
Logical. If TRUE, outputs are returned for each individual MCMC chain.

Value

A detailed summary, printed to the R console.

See also

Author

Phil J. Bouchet

Examples

if (FALSE) {
library(espresso)

# Import the example data, excluding species with sample sizes < 5
# and considering the sonar covariate
mydat <- read_data(file = NULL, min.N = 5, covariates = "sonar") 
summary(mydat)

# Configure the sampler
mydat.config <- configure_rjMCMC(dat = mydat,
                                 model.select = TRUE,
                                 covariate.select = FALSE,
                                 proposal.mh = list(t.ij = 10, mu.i = 10, 
                                                    mu = 7, phi = 10, sigma = 10),
                                 proposal.rj = list(dd = 20, cov = 7),
                                 prior.covariates = c(0, 30),
                                 n.rep = 100)
summary(mydat.config)

# Run the reversible jump MCMC
rj <- run_rjMCMC(dat = mydat.config,
                 n.chains = 2,
                 n.burn = 100,
                 n.iter = 100,
                 do.update = FALSE)
summary(rj)
}